biteco/requirements.txt
Django==5.2.5
psycopg2-binary==2.9.10
requests==2.32.3
Faker==37.1.0
python-dotenv==1.0.1